[refactor] TabBar moved to containers folder for better catergorization purposes
